Restore tree repository traits
Based in:

https://github.com/doctrine-extensions/DoctrineExtensions/tree/archive/master-2019-03-17
